Moved cli to it's own dir
