+++ cli/cli.js
@@ -10,9 +10,9 @@
 'use strict'
 var fs = require('fs');
 var program = require('commander');
-var HostManager = require('./HostManager.js');
-var SysManager = require('./SysManager.js');
-var FlowManager = require('./FlowManager.js');
+var HostManager = require('../net2/HostManager.js');
+var SysManager = require('../net2/SysManager.js');
+var FlowManager = require('../net2/FlowManager.js');
 var flowManager = new FlowManager('error');
 
 program.version('0.0.2')
@@ -124,7 +124,7 @@ console.log("Mutlicast Test", sysmanager.isMulticastIP("223.0.0.1"));
 
 var watcher = new HostManager("cli", 'client', 'error');
 
-let c = require('./MessageBus.js');
+let c = require('../net2/MessageBus.js');
 this.subscriber = new c('debug');
 
 this.subscriber.subscribe("DiscoveryEvent", "DiscoveryStart", null, (channel, ip, msg) => {
